Customer Care Call Summary/requirements.txt
ffmpeg-python
openai-whisper
gradio
langchain
langchain-openai
langchain-community
openai